Avoid "ifconfig|grep HWaddr" to find the MAC of a device
It is best to not depend on ifconfig, which output is not always the same
in different releases.

This rewrite uses /sys/class/net/<if>/address to find the MAC of a device,
in the different place where ifconfig was used.

The code is also factored out when possible.

(+ some minor cleanups)