VRF fallback feature
Allow a fallback MAC to be provided for an IP VPN attachment.

Only supported for OVS at the moment.

If a packet does not match any VRF route, its source/destination MAC
will be rewritten to the specified fallback src/dst MAC addresses,
and the packet will be sent to the specified fallback port.

This change also makes the ARP proxy in OVS IP VPN driver optional
(defaulting to False).