Really call driver.initialize() after driver.reset_state()
Commit 78f3dd25cdba37ed3a8411d53597066f790c091e had changed
the flow by calling initialize() right after __init__, but
this breaks the use case where things are do in initialize()
that must not be cleaned up by reset_state().

With this commit, initialize() is not called anymore before
reset_state().

initialize() is still guaranteed to be called before vif plugged
events and before the VPN instance event processing loop is started.